Add files via upload
Railway-Ticketing-System (C++)
Disclaimer: This was a Basic Entry Level Project for DSA, based on C++, which I worked on in my Third Semester in Bahria University Karachi Campus.

In this project of Railway Booking Management System, we will be implementing the ability to reserve and modify railway tickets.

We will use:

Linked List Operations

Searching Algorithms

Sorting Algorithms

in order to modify and print the data according to user’s needs.

This is a Basic Implementation of C++ and Data Structure & Algorithm Concepts.

We have 2 modules in this scenario, where a user or the passenger, has their own options while admin has to login separately before having their own options. It works from the entering of user details, to payment, to printing of tickets and other operations are available as well.

For Further Details About the Project, Please Refer to the Project Report File Attached.